aether_core/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# SuperAGI integration
superagi>=0.0.10

# Communication
kafka-python>=2.0.2
pika>=1.3.2
redis>=4.5.5

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6

# Utilities
python-dateutil>=2.8.2
uuid>=1.30
requests>=2.28.2
aiohttp>=3.8.4

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Documentation
pydantic-openapi-schema>=1.0.0